# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation",
        "title": "Swagger Modular FX Fiber API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PhongTran",
            "email": "phongtran11.tt@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate a user and return tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginSuccessResponseDTO"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidate user tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "post": {
                "description": "Refresh access token using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshTokenDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshTokenSuccessResponseDTO"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user and return tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterSuccessResponseDTO"
                        }
                    }
                }
            }
        },
        "/auth/register/verify-email": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verify user email address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify email",
                "parameters": [
                    {
                        "description": "Verification code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.VerifyEmailDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "List users with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ListUsersSuccessResponseDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserSuccessResponseDTO"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "description": "Get the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.GetMeSuccessResponseDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.LoginDTO": {
            "description": "Login credentials",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "secureP@ssw0rd"
                }
            }
        },
        "auth.LoginSuccessResponseDTO": {
            "description": "Response structure for successful login requests",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.TokenResponseDTO"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "auth.RefreshTokenDTO": {
            "description": "Refresh token request data",
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "auth.RefreshTokenSuccessResponseDTO": {
            "description": "Response structure for successful token refresh requests",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.TokenResponseDTO"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "auth.RegisterDTO": {
            "description": "Registration data",
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "gender": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "secureP@ssw0rd"
                },
                "phone_number": {
                    "type": "string",
                    "example": "0975234412"
                }
            }
        },
        "auth.RegisterSuccessResponseDTO": {
            "description": "Response structure for successful registration requests",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.TokenResponseDTO"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "auth.TokenResponseDTO": {
            "description": "Token response data",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "description": "in seconds",
                    "type": "integer",
                    "example": 3600
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "auth.VerifyEmailDTO": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6,
                    "example": "123456"
                }
            }
        },
        "models.UserResponseDTO": {
            "description": "User information returned in API responses",
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-01"
                },
                "deleted_at": {
                    "type": "string",
                    "example": "2023-01-10T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "email_verified": {
                    "type": "boolean",
                    "example": true
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "gender": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_login_at": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+12125551234"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T12:34:56Z"
                },
                "verify_email_code": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "user.CreateUserDTO": {
            "description": "Data for creating a new user",
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "gender": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "secureP@ssw0rd"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+12125551234"
                }
            }
        },
        "user.CreateUserSuccessResponseDTO": {
            "description": "Response structure for successful user creation requests",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.UserResponseDTO"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "user.GetMeSuccessResponseDTO": {
            "description": "Response structure for successful get me requests",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.UserResponseDTO"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "user.ListUsersSuccessResponseDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.PaginatedUsersResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "user.PaginatedUsersResponse": {
            "description": "Paginated list of users",
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserResponseDTO"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total_count": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}